refactor(core): Move cubits and update imports
Move cubits from features to core and update related
imports in home screen widgets. This improves project
structure and maintains functionality.
