+++ lib/presentation/cubits/categories/cubit/categories_state.dart
