+++ lib/presentation/cubits/favorites/cubit/favorites_cubit.dart
