+++ lib/presentation/cubits/favorites/cubit/favorites_state.dart
