+++ lib/presentation/cubits/offers/cubit/offers_state.dart
