+++ lib/presentation/screens/home/home_screen.dart
@@ -5,17 +5,17 @@ import '../../../domain/repositories/category_repository.dart';
 import '../../../domain/repositories/offers_repository.dart';
 import '../../../domain/repositories/product_repository.dart';
 import '../../../domain/repositories/store_repository.dart';
-import '../../features/cart/cubit/cart_cubit.dart';
-import '../../features/categories/cubit/categories_cubit.dart';
-import '../../features/categories/cubit/categories_state.dart';
-import '../../features/favorites/cubit/favorites_cubit.dart';
-import '../../features/favorites/cubit/favorites_state.dart';
-import '../../features/offers/cubit/offers_cubit.dart';
-import '../../features/offers/cubit/offers_state.dart';
-import '../../features/products/cubit/products_cubit.dart';
-import '../../features/products/cubit/products_state.dart';
-import '../../features/stores/cubit/stores_cubit.dart';
-import '../../features/stores/cubit/stores_state.dart';
+import '../../cubits/cart/cubit/cart_cubit.dart';
+import '../../cubits/categories/cubit/categories_cubit.dart';
+import '../../cubits/categories/cubit/categories_state.dart';
+import '../../cubits/favorites/cubit/favorites_cubit.dart';
+import '../../cubits/favorites/cubit/favorites_state.dart';
+import '../../cubits/offers/cubit/offers_cubit.dart';
+import '../../cubits/offers/cubit/offers_state.dart';
+import '../../cubits/products/cubit/products_cubit.dart';
+import '../../cubits/products/cubit/products_state.dart';
+import '../../cubits/stores/cubit/stores_cubit.dart';
+import '../../cubits/stores/cubit/stores_state.dart';
 import '../store_details/store_details_screen.dart';
 import 'widgets/home_app_bar.dart';
 import 'widgets/category_tabs.dart';
