+++ lib/presentation/screens/home/widgets/category_tabs.dart
@@ -1,8 +1,8 @@
 import 'package:flutter/material.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 import '../../../../core/design_system/widgets/chip_category.dart';
-import '../../../features/categories/cubit/categories_cubit.dart';
-import '../../../features/categories/cubit/categories_state.dart';
+import '../../../cubits/categories/cubit/categories_cubit.dart';
+import '../../../cubits/categories/cubit/categories_state.dart';
 
 
 class CategoryTabs extends StatelessWidget {
